promote aud to user level
In some cases access to the standard claim's aud can be useful for
consumer. For example to reject tokens made for different, but allowed
aud passed verification just fine. However, the full claim is not avail,
just User part of it. The change sets Audience field from std claim aud.

In other words the goal is to prevent valid token issued for subsystem A
to be used in subsystem B. The package doesn't provide such a rejection
but allow user to handle it on application level.