support both personal and work/school Microsoft accounts
The tenant for microsoft.AzureADEndpoint is currently "consumers".  As per
the
[documentation](https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-v2-protocols#endpoints),
this allows only users with personal Microsoft accounts (MSA) to sign
into the application.  Change it to "common" in order to allow users
with both personal Microsoft accounts and work/school accounts from
Azure AD to sign into the application.

Fixes #106