Fix token validation and update README
golang-jwt library is trying to validate iat claim of the ID token and
due to not accounting for clock skew, validation pretty randomly fails.

There is an open issue https://github.com/golang-jwt/jwt/issues/98 and
seems like that is fixed in v4. However it is still unclear why iat is
validation in the first place, that's not required by RFC and doesn't
seem like the right thing to do. Only nbf and exp claims should be used
for token lifetime validity check.

Also, update README to show how to configure OpenID providers.