Add custom error HTTP handlers to middleware
This patch gives ability to configure the custom HTTP handlers,
which render authentication and authorization errors.
In some use-cases it is necessary to produce different
Content-Type, body, HTTP headers and status codes
than ones provided by default.
It is possible to configure one global handler,
which is same for all web routes in the Authenticator,
and multiple on-demand handlers
for different middleware instances and web routes.